basic compile pass

+++ packages/api-server/src/utils.ts
@@ -6,7 +6,7 @@ import {
   stringifyError,
 } from "@dendronhq/common-all";
 import { createLogger } from "@dendronhq/common-server";
-import execa, { ExecaChildProcess } from "execa";
+import { Subprocess } from "execa";
 import _ from "lodash";
 import path from "path";
 import { launchv2 } from ".";
@@ -61,7 +61,7 @@ export class ServerUtils {
     subprocess,
     cb,
   }: {
-    subprocess: ExecaChildProcess;
+    subprocess: Subprocess;
     cb: (exitType: SubProcessExitType, args?: any) => any;
   }) {
     subprocess.on("exit", () => cb(SubProcessExitType.EXIT));
@@ -80,11 +80,13 @@ export class ServerUtils {
    * Attach to a server process to kill it when the current process exits
    * @param subprocess
    */
-  static cleanServerProcess(subprocess: ExecaChildProcess) {
+  static cleanServerProcess(subprocess: Subprocess) {
     const handleExit = () => {
       console.log("handle exit");
       try {
-        process.kill(subprocess.pid);
+        if (subprocess.pid) {
+          process.kill(subprocess.pid);
+        }
       } catch (err: any) {
         // this means process was already killed
         if (err.code !== "ESRCH") {
@@ -174,14 +176,15 @@ export class ServerUtils {
     port,
     googleOauthClientId,
     googleOauthClientSecret,
-  }: ServerArgs): Promise<{ port: number; subprocess: ExecaChildProcess }> {
+  }: ServerArgs): Promise<{ port: number; subprocess: Subprocess }> {
     const logger = createLogger(
       "execServer",
       path.join(logPath, "dendron.log")
     );
+    const execa = await import("execa");
     return new Promise((resolve, reject) => {
       logger.info({ state: "enter" });
-      const subprocess = execa.node(scriptPath, {
+      const subprocess = execa.execaNode(scriptPath, {
         env: {
           LOG_PATH: logPath,
           ENGINE_SERVER_PORT: port,
@@ -193,7 +196,7 @@ export class ServerUtils {
         } as SERVER_ENV,
       });
       logger.info({ state: "post:exec.node" });
-      subprocess.on("close", (code) => {
+      subprocess.on("close", (code: any) => {
         logger.error({ state: "close" });
         reject(new DendronError({ message: "close", payload: { code } }));
       });
@@ -201,17 +204,17 @@ export class ServerUtils {
         logger.error({ state: "disconnect" });
         reject(new DendronError({ message: "disconnect" }));
       });
-      subprocess.on("exit", (code) => {
+      subprocess.on("exit", (code: any) => {
         logger.error({ state: "exit" });
         reject(new DendronError({ message: "exit", payload: { code } }));
       });
-      subprocess.on("error", (err) => {
+      subprocess.on("error", (err: Error) => {
         logger.error({ state: "error", payload: err });
         reject(
           new DendronError({ message: "error", payload: stringifyError(err) })
         );
       });
-      subprocess.on("message", (message) => {
+      subprocess.on("message", (message: any) => {
         logger.info({ state: "message", message });
         const port = parseInt(message as string, 10);
         if (port <= 0) {
